refactor(blobnode): report network flow metric with io type
with: #22236519 of #22160501
with: #22237444 of #22160501
